fix: resolve all ESLint warnings
- Fix matoran.ts parsing error: add missing 'Kivi' property name
- Add coverage to ESLint ignores (generated files)
- Fix MatoranAvatar useMemo: add matoran to dependency array
- Fix Tooltip: wrap updatePosition in useCallback, add to effect deps
- Fix useNuvaMask: add maskColor to first useEffect dependency array
- Fix Settings react-refresh: split useSettings and context into separate files
