+++ src/context/Settings.tsx
@@ -1,4 +1,4 @@
-import { createContext, useContext, useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import {
   getDebugMode,
   getShadowsEnabled,
@@ -6,20 +6,7 @@ import {
   saveShadowsEnabled,
 } from '../services/gamePersistence';
 
-export interface SettingsState {
-  debugMode: boolean;
-  setDebugMode: (value: boolean) => void;
-  shadowsEnabled: boolean;
-  setShadowsEnabled: (value: boolean) => void;
-}
-
-const SettingsContext = createContext<SettingsState | null>(null);
-
-export function useSettings() {
-  const ctx = useContext(SettingsContext);
-  if (!ctx) throw new Error('useSettings must be used within SettingsProvider');
-  return ctx;
-}
+import { SettingsContext } from './SettingsContext';
 
 export function SettingsProvider({ children }: { children: React.ReactNode }) {
   const [debugMode, setDebugModeState] = useState(getDebugMode);
